Move color picker in its own component
